+++ src/test/java/persistence/sql/ddl/h2/meta/NullableTest.java
@@ -4,7 +4,7 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Id;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import persistence.sql.dialect.H2Dialect;
+import persistence.sql.ddl.dialect.H2Dialect;
 
 import java.lang.reflect.Field;
